Copy changes from patchfile into a fork
* disables MathLive's `smartFence` feature in certain situations in
which it is applied regardless of the provided `smartFence` setting. This
fixes a glitch that occurs when an opening delimiter is placed at the
beginning of the input.

* prevents 'math-json' serialization when copying to the clipboard (and
deserialization when pasting). This is a custom format that the author is
still working out; notably, it fails to properly serialize intergrals.

* prevents the mathfield from closing when blurred.

* adds breakpoint CSS classes to the virtual keyboard container in order to
allow responsive styling with respect to the container.

* restores native mobile/touch keyboard functionality.